+++ soxspipe/recipes/tests/test_soxs_mdark.py
@@ -53,7 +53,8 @@ class test_soxs_mdark(unittest.TestCase):
             settings=settings,
             inputFrames=sofPath
         )
-        this.produce_product()
+        productPath = this.produce_product()
+        print(f"Here is the final product `{productPath}`")
 
         sofPath = "~/xshooter-pipeline-data/unittest_data/xshooter-darks/sofs/nir_240s_darks.sof"
         from soxspipe.recipes import soxs_mdark
@@ -62,7 +63,8 @@ class test_soxs_mdark(unittest.TestCase):
             settings=settings,
             inputFrames=sofPath
         )
-        this.produce_product()
+        productPath = this.produce_product()
+        print(f"Here is the final product `{productPath}`")
 
         sofPath = "~/xshooter-pipeline-data/unittest_data/xshooter-darks/sofs/nir_125s_darks.sof"
         from soxspipe.recipes import soxs_mdark
@@ -71,7 +73,8 @@ class test_soxs_mdark(unittest.TestCase):
             settings=settings,
             inputFrames=sofPath
         )
-        this.produce_product()
+        productPath = this.produce_product()
+        print(f"Here is the final product `{productPath}`")
 
     def test_soxs_mdark_uvb_function(self):
 
@@ -82,7 +85,8 @@ class test_soxs_mdark(unittest.TestCase):
             settings=settings,
             inputFrames=sofPath
         )
-        this.produce_product()
+        productPath = this.produce_product()
+        print(f"Here is the final product `{productPath}`")
 
         sofPath = "~/xshooter-pipeline-data/unittest_data/xshooter-darks/sofs/uvb_2x2_dark_3600s.sof"
         from soxspipe.recipes import soxs_mdark
@@ -91,7 +95,8 @@ class test_soxs_mdark(unittest.TestCase):
             settings=settings,
             inputFrames=sofPath
         )
-        this.produce_product()
+        productPath = this.produce_product()
+        print(f"Here is the final product `{productPath}`")
 
     def test_soxs_mdark_vis_function(self):
         sofPath = "~/xshooter-pipeline-data/unittest_data/xshooter-darks/sofs/vis_1x1_dark_3600s.sof"
@@ -101,7 +106,8 @@ class test_soxs_mdark(unittest.TestCase):
             settings=settings,
             inputFrames=sofPath
         )
-        this.produce_product()
+        productPath = this.produce_product()
+        print(f"Here is the final product `{productPath}`")
 
         sofPath = "~/xshooter-pipeline-data/unittest_data/xshooter-darks/sofs/vis_2x2_dark_3600s.sof"
         from soxspipe.recipes import soxs_mdark
@@ -110,7 +116,8 @@ class test_soxs_mdark(unittest.TestCase):
             settings=settings,
             inputFrames=sofPath
         )
-        this.produce_product()
+        productPath = this.produce_product()
+        print(f"Here is the final product `{productPath}`")
 
     # def test_soxs_mdark_function(self):